Bug 29233: Correct missed jQueryUI datepicker in serials search sidebar
This patch fixes the serials search sidebar so that the "Expires before"
field uses Flatpickr instead of jQueryUI datepicker.

To test, apply the patch and go to Serials. Perform a subscription
search which will return multiple results.

In the sidebar, test that the "Expires before" field uses a
Flatpickr widget and works correctly.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>