Bug 29430: Table cell click doesn't activate buttons in patron search
Bug 15812 included a change which allows a click on the patron
search results table cell to toggle the checkbox it contains. This patch
modifies that click event so that it fires the change() event which is
required for toggling the "Add to patron list" and "Merge patrons"
buttons.

To reproduce this problem, perform a patron search in the staff client
which will return multiple results.
 - In the first column containing checkboxes, click in the empty part of
   the table cell. The checkbox should be checked.
   - However, the "Add to patron list" button remains disabled.
  - Clicking a table cell to check another checkbox should result in the
    "Merge selected patrons" button being enabled, but it doesn't.

To test, apply the patch and repeat the process above. The behavior of
the buttons should be the same whether you're clicking the checkbox
itself or the table cell it's in.

Signed-off-by: Barbara Johnson <barbara.johnson@bedfordtx.gov>

Signed-off-by: Joonas Kylmälä <joonas.kylmala@iki.fi>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>